assigment1/requirements.txt
absl-py==2.3.1
altair==5.5.0
anaconda==0.0.1.1
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
autopep8==1.6.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.2
cachetools==6.2.2
certifi==2025.8.3
cffi==1.14.5
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
csbdeep==0.8.1
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
deskew==1.5.3
Django==5.2.1
django-cors-headers==3.10.1
django-storages==1.12.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
folium==0.20.0
fonttools==4.59.1
fqdn==1.5.1
fsspec==2025.9.0
gast==0.6.0
gensim==4.3.3
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
graphviz==0.21
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
ImageIO==2.37.2
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==3.11.2
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.45.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
milksnake==0.1.5
mistune==3.1.3
ml_dtypes==0.5.3
mpmath==1.3.0
mssql-django==1.5
mypy-extensions==0.4.3
mysqlclient==2.2.7
namex==0.1.0
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numba==0.62.1
numpy==1.26.4
opencv-contrib-python==4.12.0.88
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.2
pillow==11.3.0
platformdirs==4.3.8
polars==1.35.2
polars-runtime-32==1.35.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycodestyle==2.8.0
pycparser==2.20
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
PyJWT==2.3.0
pyodbc==5.2.0
pyparsing==3.2.3
python-bidi==0.6.7
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2021.3
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.1.2
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.4.2
stack-data==0.6.3
stardist==0.9.1
statsmodels==0.14.5
streamlit==1.51.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
types-python-dateutil==2.9.0.20250809
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.232
ultralytics-thop==2.0.18
underthesea==8.3.0
underthesea_core==1.0.5
uri-template==1.3.0
urllib3==2.5.0
urlquote==1.1.4
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==5.3.0
widgetsnbextension==4.0.14
wrapt==1.17.3
xgboost==3.0.5
xyzservices==2025.4.0
